@@ -9,23 +9,26 @@ class ControllerPage extends StatelessWidget {
 
   ControllerPage(this.device);
 
-  Future<bool> _save() {
-    print('close and disconnect');
-    // MidiCommand().teardown();
-    MidiCommand().disconnectDevice(device);
-    return Future.value(true);
-  }
+  // Future<bool> _save() {
+  //   print('close and disconnect');
+  //   // MidiCommand().teardown();
+  //   MidiCommand().disconnectDevice(device);
+  //   return Future.value(true);
+  // }
 
   @override
   Widget build(BuildContext context) {
-    return WillPopScope(
-        onWillPop: _save,
-        child: Scaffold(
-          appBar: AppBar(
-            title: Text('Controls'),
-          ),
-          body: MidiControls(device),
-        ));
+    return
+        // WillPopScope(
+        //   onWillPop: _save,
+        //   child:
+        Scaffold(
+      appBar: AppBar(
+        title: Text('Controls'),
+      ),
+      body: MidiControls(device),
+      // )
+    );
   }
 }
 
@@ -46,19 +49,18 @@ class MidiControlsState extends State<MidiControls> {
   var _value = 0;
 
   // StreamSubscription<String> _setupSubscription;
-  StreamSubscription<MidiPacket> _rxSubscription;
+  StreamSubscription<MidiPacket>? _rxSubscription;
   MidiCommand _midiCommand = MidiCommand();
 
   @override
   void initState() {
     // print('init controller');
-    _rxSubscription = _midiCommand.onMidiDataReceived.listen((packet) {
+    _rxSubscription = _midiCommand.onMidiDataReceived?.listen((packet) {
       // print('received packet $packet');
       var data = packet.data;
       var timestamp = packet.timestamp;
       var device = packet.device;
-      print(
-          "data $data @ time $timestamp from device ${device.name}:${device.id}");
+      print("data $data @ time $timestamp from device ${device.name}:${device.id}");
 
       var status = data[0];
 
@@ -100,8 +102,7 @@ class MidiControlsState extends State<MidiControls> {
       child: Column(
         children: <Widget>[
           SteppedSelector('Channel', _channel + 1, 1, 16, _onChannelChanged),
-          SteppedSelector(
-              'Controller', _controller, 0, 127, _onControllerChanged),
+          SteppedSelector('Controller', _controller, 0, 127, _onControllerChanged),
           SlidingSelector('Value', _value, 0, 127, _onValueChanged),
         ],
       ),
@@ -123,8 +124,7 @@ class MidiControlsState extends State<MidiControls> {
   _onValueChanged(int newValue) {
     setState(() {
       _value = newValue;
-      CCMessage(channel: _channel, controller: _controller, value: _value)
-          .send();
+      CCMessage(channel: _channel, controller: _controller, value: _value).send();
     });
   }
 }
@@ -136,8 +136,7 @@ class SteppedSelector extends StatelessWidget {
   final int value;
   final Function(int) callback;
 
-  SteppedSelector(
-      this.label, this.value, this.minValue, this.maxValue, this.callback);
+  SteppedSelector(this.label, this.value, this.minValue, this.maxValue, this.callback);
 
   @override
   Widget build(BuildContext context) {
@@ -172,8 +171,7 @@ class SlidingSelector extends StatelessWidget {
   final int value;
   final Function(int) callback;
 
-  SlidingSelector(
-      this.label, this.value, this.minValue, this.maxValue, this.callback);
+  SlidingSelector(this.label, this.value, this.minValue, this.maxValue, this.callback);
 
   @override
   Widget build(BuildContext context) {
