@@ -12,7 +12,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  StreamSubscription<String> _setupSubscription;
+  StreamSubscription<String>? _setupSubscription;
   MidiCommand _midiCommand = MidiCommand();
 
   @override
@@ -22,15 +22,20 @@ class _MyAppState extends State<MyApp> {
     _midiCommand.startScanningForBluetoothDevices().catchError((err) {
       print("Error $err");
     });
-    _setupSubscription = _midiCommand.onMidiSetupChanged.listen((data) {
+    _setupSubscription = _midiCommand.onMidiSetupChanged?.listen((data) {
       print("setup changed $data");
 
+      setState(() {});
       switch (data) {
         case "deviceFound":
-          setState(() {});
+          print("device found");
           break;
-        // case "deviceOpened":
-        //   break;
+        case "deviceOpened":
+          print("device found");
+          break;
+        case "deviceLost":
+          print("device lost");
+          break;
         default:
           // print("Unhandled setup change: $data");
           break;
@@ -40,7 +45,7 @@ class _MyAppState extends State<MyApp> {
 
   @override
   void dispose() {
-    _setupSubscription.cancel();
+    _setupSubscription?.cancel();
     super.dispose();
   }
 
@@ -49,13 +54,11 @@ class _MyAppState extends State<MyApp> {
     return new MaterialApp(
       home: new Scaffold(
         appBar: new AppBar(
-          title: const Text('FlutterMidiCommand'),
+          title: const Text('FlutterMidiCommand Example'),
           actions: <Widget>[
             IconButton(
                 onPressed: () {
-                  _midiCommand
-                      .startScanningForBluetoothDevices()
-                      .catchError((err) {
+                  _midiCommand.startScanningForBluetoothDevices().catchError((err) {
                     print("Error $err");
                   });
                   setState(() {});
@@ -63,43 +66,55 @@ class _MyAppState extends State<MyApp> {
                 icon: Icon(Icons.refresh))
           ],
         ),
-        body: new Center(
-            child: FutureBuilder(
-                future: _midiCommand.devices,
-                builder: (BuildContext context, AsyncSnapshot snapshot) {
-                  if (snapshot.hasData && snapshot.data != null) {
-                    var devices = snapshot.data as List<MidiDevice>;
-                    return ListView.builder(
-                      // Let the ListView know how many items it needs to build
-                      itemCount: devices.length,
-                      // Provide a builder function. This is where the magic happens! We'll
-                      // convert each item into a Widget based on the type of item it is.
-                      itemBuilder: (context, index) {
-                        MidiDevice device = devices[index];
+        bottomNavigationBar: Container(
+          padding: EdgeInsets.all(24.0),
+          child: Text(
+            "Tap to connnect/disconnect, long press to control.",
+            textAlign: TextAlign.center,
+          ),
+        ),
+        body: Center(
+          child: FutureBuilder(
+            future: _midiCommand.devices,
+            builder: (BuildContext context, AsyncSnapshot snapshot) {
+              if (snapshot.hasData && snapshot.data != null) {
+                var devices = snapshot.data as List<MidiDevice>;
+                return ListView.builder(
+                  itemCount: devices.length,
+                  itemBuilder: (context, index) {
+                    MidiDevice device = devices[index];
 
-                        return ListTile(
-                          title: Text(
-                            device.name,
-                            style: Theme.of(context).textTheme.headline,
-                          ),
-                          subtitle: Text(
-                              "ins:${device.inputPorts.length} outs:${device.outputPorts.length}"),
-                          trailing: device.type == "BLE"
-                              ? Icon(Icons.bluetooth)
-                              : null,
-                          onTap: () {
-                            _midiCommand.connectToDevice(device);
-                            Navigator.of(context).push(MaterialPageRoute<Null>(
-                              builder: (_) => ControllerPage(device),
-                            ));
-                          },
-                        );
+                    return ListTile(
+                      title: Text(
+                        device.name,
+                        style: Theme.of(context).textTheme.headline5,
+                      ),
+                      subtitle: Text("ins:${device.inputPorts.length} outs:${device.outputPorts.length}"),
+                      leading: Icon(device.connected ? Icons.radio_button_on : Icons.radio_button_off),
+                      trailing: device.type == "BLE" ? Icon(Icons.bluetooth) : null,
+                      onLongPress: () {
+                        Navigator.of(context).push(MaterialPageRoute<Null>(
+                          builder: (_) => ControllerPage(device),
+                        ));
+                      },
+                      onTap: () {
+                        if (device.connected) {
+                          print("disconnect");
+                          _midiCommand.disconnectDevice(device);
+                        } else {
+                          print("connect");
+                          _midiCommand.connectToDevice(device);
+                        }
                       },
                     );
-                  } else {
-                    return new CircularProgressIndicator();
-                  }
-                })),
+                  },
+                );
+              } else {
+                return new CircularProgressIndicator();
+              }
+            },
+          ),
+        ),
       ),
     );
   }
